Design of the pages starting
