+++ packages/nextjs/app/about-us/page.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+const AboutUsPage = () => {
+  return (
+    <div className="text-white text-center py-20">
+      <h1 className="text-4xl font-bold">About Us</h1>
+      <p className="mt-4 text-lg">
+        Welcome to StarkLotto! We are the future of decentralized lotteries.
+      </p>
+    </div>
+  );
+};
+
+export default AboutUsPage;
