+++ packages/nextjs/app/contact-us/page.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const ContactUsPage = () => {
+  return (
+    <div className="text-white text-center py-20">
+      <h1 className="text-4xl font-bold">Contact Us</h1>
+      <p className="mt-4 text-lg">Have any questions or feedback?</p>
+    </div>
+  );
+};
+
+export default ContactUsPage;
