+++ packages/nextjs/app/how-it-works/page.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+const HowItWorksPage = () => {
+  return (
+    <div className="text-white text-center py-20">
+      <h1 className="text-4xl font-bold">How It Works</h1>
+      <p className="mt-4 text-lg">
+        Learn how to participate and win amazing prizes with StarkLotto!
+      </p>
+    </div>
+  );
+};
+
+export default HowItWorksPage;
