+++ packages/nextjs/app/principal/page.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const PlayPage = () => {
+  return (
+    <div className="text-white text-center py-20 mt-24">
+      <h1 className="text-4xl font-bold">Holaaa</h1>
+    </div>
+  );
+};
+
+export default PlayPage;
