+++ packages/nextjs/app/prizes/page.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+const PrizesPage = () => {
+  return (
+    <div className="text-white text-center py-20">
+      <h1 className="text-4xl font-bold">Prizes</h1>
+      <p className="mt-4 text-lg">
+        Discover the amazing prizes you can win with StarkLotto!
+      </p>
+    </div>
+  );
+};
+
+export default PrizesPage;
