+++ packages/nextjs/components/Header.tsx
@@ -1,186 +1,64 @@
 "use client";
 
-import React, { useCallback, useRef, useState, useEffect } from "react";
-import Image from "next/image";
+import React from "react";
 import Link from "next/link";
-import { usePathname } from "next/navigation";
-import { Bars3Icon, BugAntIcon } from "@heroicons/react/24/outline";
-import { useOutsideClick } from "~~/hooks/scaffold-stark";
-import { CustomConnectButton } from "~~/components/scaffold-stark/CustomConnectButton";
-import { useTheme } from "next-themes";
-import { useTargetNetwork } from "~~/hooks/scaffold-stark/useTargetNetwork";
-import { devnet } from "@starknet-react/chains";
-import { SwitchTheme } from "./SwitchTheme";
-import { useAccount, useNetwork, useProvider } from "@starknet-react/core";
-import { BlockIdentifier } from "starknet";
+import Image from "next/image";
+import { WalletIcon, UserIcon } from "@heroicons/react/24/outline";
 
-type HeaderMenuLink = {
-  label: string;
-  href: string;
-  icon?: React.ReactNode;
-};
-
-export const menuLinks: HeaderMenuLink[] = [
-  {
-    label: "Home",
-    href: "/",
-  },
-  {
-    label: "Debug Contracts",
-    href: "/debug",
-    icon: <BugAntIcon className="h-4 w-4" />,
-  },
+const menuLinks = [
+  { label: "Home", href: "/" },
+  { label: "How It Works", href: "/how-it-works" },
+  { label: "Play", href: "/play" },
+  { label: "Prizes", href: "/prizes" },
+  { label: "Contact Us", href: "/contact-us" },
+  { label: "About Us", href: "/about-us" },
 ];
 
-export const HeaderMenuLinks = () => {
-  const pathname = usePathname();
-  const { theme } = useTheme();
-  const [isDark, setIsDark] = useState(false);
-
-  useEffect(() => {
-    setIsDark(theme === "dark");
-  }, [theme]);
-  return (
-    <>
-      {menuLinks.map(({ label, href, icon }) => {
-        const isActive = pathname === href;
-        return (
-          <li key={href}>
-            <Link
-              href={href}
-              passHref
-              className={`${
-                isActive
-                  ? "!bg-gradient-nav !text-white active:bg-gradient-nav shadow-md"
-                  : ""
-              } py-1.5 px-3 text-sm rounded-full gap-2 grid grid-flow-col hover:bg-gradient-nav hover:text-white`}
-            >
-              {icon}
-              <span>{label}</span>
-            </Link>
-          </li>
-        );
-      })}
-    </>
-  );
-};
-
-/**
- * Site header
- */
 export const Header = () => {
-  const [isDrawerOpen, setIsDrawerOpen] = useState(false);
-  const burgerMenuRef = useRef<HTMLDivElement>(null);
-
-  useOutsideClick(
-    //@ts-expect-error refs are supposed to be null by default
-    burgerMenuRef,
-    useCallback(() => setIsDrawerOpen(false), []),
-  );
-
-  const { targetNetwork } = useTargetNetwork();
-  const isLocalNetwork = targetNetwork.network === devnet.network;
-
-  const { provider } = useProvider();
-  const { address, status, chainId } = useAccount();
-  const { chain } = useNetwork();
-  const [isDeployed, setIsDeployed] = useState(true);
-
-  useEffect(() => {
-    if (
-      status === "connected" &&
-      address &&
-      chainId === targetNetwork.id &&
-      chain.network === targetNetwork.network
-    ) {
-      provider
-        .getClassHashAt(address)
-        .then((classHash) => {
-          if (classHash) setIsDeployed(true);
-          else setIsDeployed(false);
-        })
-        .catch((e) => {
-          console.error("contreact cehc", e);
-          if (e.toString().includes("Contract not found")) {
-            setIsDeployed(false);
-          }
-        });
-    }
-  }, [
-    status,
-    address,
-    provider,
-    chainId,
-    targetNetwork.id,
-    targetNetwork.network,
-    chain.network,
-  ]);
-
   return (
-    <div className=" lg:static top-0 navbar min-h-0 flex-shrink-0 justify-between z-20 px-0 sm:px-2">
-      <div className="navbar-start w-auto lg:w-1/2 -mr-2">
-        <div className="lg:hidden dropdown" ref={burgerMenuRef}>
-          <label
-            tabIndex={0}
-            className={`ml-1 btn btn-ghost 
-              [@media(max-width:379px)]:!px-3 [@media(max-width:379px)]:!py-1 
-              [@media(max-width:379px)]:!h-9 [@media(max-width:379px)]:!min-h-0
-              [@media(max-width:379px)]:!w-10
-              ${isDrawerOpen ? "hover:bg-secondary" : "hover:bg-transparent"}`}
-            onClick={() => {
-              setIsDrawerOpen((prevIsOpenState) => !prevIsOpenState);
-            }}
-          >
-            <Bars3Icon className="h-1/2" />
-          </label>
-          {isDrawerOpen && (
-            <ul
-              tabIndex={0}
-              className="menu menu-compact dropdown-content mt-3 p-2 shadow rounded-box w-52 bg-base-100"
-              onClick={() => {
-                setIsDrawerOpen(false);
-              }}
-            >
-              <HeaderMenuLinks />
-            </ul>
-          )}
-        </div>
-        <Link
-          href="/"
-          passHref
-          className="hidden lg:flex items-center gap-2 ml-4 mr-6 shrink-0"
-        >
-          <div className="flex relative w-10 h-10">
+    <header className="w-full bg-transparent fixed top-0 z-50">
+      <nav className="container mx-auto flex items-center justify-between py-4 px-8">
+        {/* Logo */}
+        <Link href="/" passHref>
+          <div className="flex items-center gap-2 cursor-pointer">
             <Image
-              alt="SE2 logo"
-              className="cursor-pointer"
-              fill
-              src="/logo.svg"
+              src="/logo.svg" // Asegúrate de tener tu logo en la carpeta /public
+              alt="StarkLotto Logo"
+              width={40}
+              height={40}
+              className="rounded-full"
             />
-          </div>
-          <div className="flex flex-col">
-            <span className="font-bold leading-tight">Scaffold-Stark</span>
-            <span className="text-xs">Starknet dev stack</span>
+            <span className="text-white text-lg font-bold">STARKLOTTO</span>
           </div>
         </Link>
-        <ul className="hidden lg:flex lg:flex-nowrap menu menu-horizontal px-1 gap-2">
-          <HeaderMenuLinks />
+
+        {/* Menú de Navegación */}
+        <ul className="hidden md:flex space-x-8">
+          {menuLinks.map(({ label, href }) => (
+            <li key={href}>
+              <Link
+                href={href}
+                passHref
+                className="text-white hover:text-yellow-400 transition duration-300"
+              >
+                {label}
+              </Link>
+            </li>
+          ))}
         </ul>
-      </div>
-      <div className="navbar-end flex-grow mr-2 gap-4">
-        {status === "connected" && !isDeployed ? (
-          <span className="bg-[#8a45fc] text-[9px] p-1 text-white">
-            Wallet Not Deployed
-          </span>
-        ) : null}
-        <CustomConnectButton />
-        {/* <FaucetButton /> */}
-        <SwitchTheme
-          className={`pointer-events-auto ${
-            isLocalNetwork ? "mb-1 lg:mb-0" : ""
-          }`}
-        />
-      </div>
-    </div>
+
+        {/* Iconos */}
+        <div className="flex items-center space-x-4">
+          <button className="p-2 bg-transparent hover:bg-gray-700 rounded-full transition">
+            <WalletIcon className="h-6 w-6 text-white" />
+          </button>
+          <button className="p-2 bg-transparent hover:bg-gray-700 rounded-full transition">
+            <UserIcon className="h-6 w-6 text-white" />
+          </button>
+        </div>
+      </nav>
+    </header>
   );
 };
+
+export default Header;
