+++ packages/nextjs/styles/globals.css
@@ -8,7 +8,15 @@
 }
 
 body {
+  margin: 0;
+  padding: 0;
   min-height: 100vh;
+  background: linear-gradient(135deg, #3a0909 0%, #000000 100%);
+  background-attachment: fixed;
+  background-size: cover;
+  font-family: "Inter", sans-serif;
+  color: white;
+  overflow-x: hidden;
 }
 
 h1,
